New Updated on Repository. Brief change on Entities.

+++ Entity/src/main/java/com/itlize/firstProject/Repository/Project_ResourcesRepository.java
@@ -4,7 +4,16 @@ import com.itlize.firstProject.Entity.Project_Resources;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface Project_ResourcesRepository extends JpaRepository<Project_Resources, Integer> {
+public interface Project_ResourcesRepository extends JpaRepository<Project_Resources, Long> {
+
+
+
 }
+
+
+
+
